Storyboard Feed Layout (#20)
* feat: add Feed storyboard and assets with layout from prototype

* refactor: instantiate `FeedViewController` from feed storyboard

NOTE: Notice how we must use property injection to configure a view that is initialized from a storyboard

* refactor: move `FeedViewController` + `FeedRefreshViewController` composition (instantiation and configuration) to the Storyboard

* Merge `FeedRefreshViewController` with `FeedViewController` since the `UIRefreshControl` creation and configuration now lives in the Storyboard

* refactor: move tableView prefetchDataSource setup to the storyboard

* feat: reuse table view cells from storyboard configuration

* refactor: Add UITableView helper to dequeue typed celss by class name

* feat: display image with animation

* fix: iOS 15 lifecycle changes where reusing a cell that was never displayed can cause the the wrong image to load at the wrong index path

* fix: ensure previous cell prepareForReuse does not affect new cells by reumoving the `onReuse` closure reference when releasing the cell for reuse.

REASON: Every time we update the table view state, existing cells can be reused for different index paths, where they'll be managed by different cell controllers - so we need to remove all references to the previous cell controllers when releasing the cell for reuse. Since the `onReuse` closure references the cell controller via `self`, we need to set `onReuse` to `nil` when releasing the cell for reuse by another cell controller. This way, we ensure there are no references to the cell controller so there will be no side effects.